# Backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Google Forms Quiz System","version":"0.1.0"},"paths":{"/__health-check__":{"get":{"summary":"Health Check","description":"Endpoint to verify API and service account functionality.","operationId":"health_check___health_check___get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/quizzes/{quiz_id}":{"get":{"summary":"Get Quiz","description":"Get details for a specific quiz.","operationId":"get_quiz_quizzes__quiz_id__get","parameters":[{"name":"quiz_id","in":"path","required":true,"schema":{"type":"string","title":"Quiz Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Quiz","operationId":"delete_quiz_quizzes__quiz_id__delete","parameters":[{"name":"quiz_id","in":"path","required":true,"schema":{"type":"string","title":"Quiz Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/quizzes/":{"get":{"summary":"List Quizzes","operationId":"list_quizzes_quizzes__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/QuizDB"},"type":"array","title":"Response List Quizzes Quizzes  Get"}}}}}},"post":{"summary":"Create Quiz","description":"Takes questions, creates a Google Form, saves to MongoDB as DRAFT.","operationId":"create_quiz_quizzes__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/quizzes/{quiz_id}/approve":{"post":{"summary":"Approve Quiz","description":"Transition: Draft -> Approved. Triggers email and updates DB.","operationId":"approve_quiz_quizzes__quiz_id__approve_post","parameters":[{"name":"quiz_id","in":"path","required":true,"schema":{"type":"string","title":"Quiz Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApproveRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ApproveRequest":{"properties":{"emails":{"items":{"type":"string","format":"email"},"type":"array","title":"Emails"}},"type":"object","required":["emails"],"title":"ApproveRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Question":{"properties":{"text":{"type":"string","title":"Text"},"options":{"items":{"type":"string"},"type":"array","title":"Options"},"correct_answer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Correct Answer"}},"type":"object","required":["text","options"],"title":"Question"},"QuizCreateRequest":{"properties":{"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","default":"Generated via Quiz System"},"questions":{"items":{"$ref":"#/components/schemas/Question"},"type":"array","title":"Questions"}},"type":"object","required":["title","questions"],"title":"QuizCreateRequest"},"QuizDB":{"properties":{"_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"title":{"type":"string","title":"Title"},"status":{"type":"string","title":"Status","default":"draft"},"form_url":{"type":"string","title":"Form Url"},"form_id":{"type":"string","title":"Form Id"},"questions":{"items":{"$ref":"#/components/schemas/Question"},"type":"array","title":"Questions"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["title","form_url","form_id","questions"],"title":"QuizDB","example":{"form_id":"12345","form_url":"https://docs.google.com/forms/...","questions":[],"status":"draft","title":"Math Quiz"}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}